core/logger: group critical code together; add test
