+++ core/logger/logger_test.go
@@ -6,8 +6,14 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestNoSampling(t *testing.T) {
+func TestConfig(t *testing.T) {
+	// no sampling
 	assert.Nil(t, newBaseConfig().Sampling)
 	assert.Nil(t, newTestConfig().Sampling)
 	assert.Nil(t, newProductionConfig("", false, true, false).Sampling)
+
+	// not development, which would trigger panics for Critical level
+	assert.False(t, newBaseConfig().Development)
+	assert.False(t, newTestConfig().Development)
+	assert.False(t, newProductionConfig("", false, true, false).Development)
 }
